+++ CHANGELOG.md
@@ -5,10 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## Unreleased
+
+### Added
+- Added support for Product Types if Commerce is installed
+
+
 ## 1.0.4 - 2018-09-26
 
 ### Fixed
-- Fixed an error in composer.json where the version string was not matching the tag name.
+- Fixed an error in composer.json where the version string was not matching the tag name
 
 
 ## 1.0.3 - 2018-09-25